+++ packages/node/src/index.ts
@@ -2,11 +2,10 @@
 
 // Re-export useful types from core
 export type {
-  ExplicitJourneyStage,
   IngestResponse,
   ServerIdentifyOptions,
   ServerTrackOptions,
   TrackerConfig,
 } from "@outlit/core"
-export type { BillingOptions, OutlitOptions, StageOptions, UserMethods } from "./client"
+export type { OutlitOptions, UserMethods } from "./client"
 export { Outlit } from "./client"